requirements.txt
pydantic
openai
fastapi
python-dotenv